Дробит main.js на модули
